# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MOA Prototype API",
    "version": "0.1.0",
    "description": "Demo authentication and payments only. Not a production PG or customs API."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/offers/{id}/accept": {
      "post": {
        "summary": "선결제된 부탁에서 한 명 선택 · PAYMENT_HELD 거래 생성 · 추가 청구 없음",
        "operationId": "post_offers_id_accept",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptOffer"
              }
            }
          }
        }
      }
    },
...
    "/transactions/{id}/actions": {
      "post": {
        "summary": "권한과 버전 검증 상태 명령",
        "operationId": "post_transactions_id_actions",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionAction"
              }
            }
          }
        }
      }
    },
    "/transactions/{id}/reviews": {
      "post": {
        "summary": "구매 확정 후 후기",
        "operationId": "post_transactions_id_reviews",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque, random 24-hour demo session token"
      }
    },
    "schemas": {
...
      "AcceptOffer": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000000
          }
        },
        "required": [
          "expectedRevision"
        ]
      },
...
      "ReviewInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "text": {
            "type": "string",
            "minLength": 2,
            "maxLength": 500
          }
        },
        "required": [
          "rating",
          "text"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "Transaction": {
        "type": "object",
        "properties": {
          "prepaid": { "type": "boolean", "description": "True when funded before buyer selection; absent for legacy matches" },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "productPrice": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000000
          },
          "travelerReward": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000000
          },
          "platformFee": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000000
          },
          "shippingFee": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000000
          },
          "taxReserve": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000000
          },
          "totalPrice": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000000
          },
          "fxRate": {
            "type": "number"
          },
          "priceSource": {
            "type": "string",
            "enum": [
              "DEMO_FIXED"
            ]
          },
          "requestId": {
            "type": "string"
          },
          "offerId": {
            "type": "string"
          },
          "buyerId": {
            "type": "string"
          },
          "travelerId": {
            "type": "string"
          },
          "transport": {
            "type": "string",
            "enum": [
              "INTERNATIONAL_SHIPPING",
              "DOMESTIC_PARCEL",
              "MEETUP"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "REQUESTED",
              "OFFER_RECEIVED",
              "MATCHED",
              "PAYMENT_HELD",
              "PURCHASED",
              "TRAVELING",
              "SHIPPED",
              "DELIVERED",
              "CONFIRMED",
              "SETTLED",
              "CANCELLED",
              "DISPUTED"
            ]
          },
          "revision": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000000
          },
          "estimatedDeliveryDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "ActionTRAVEL": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "TRAVEL"
            ]
          },
          "expectedRevision": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000000
          }
        },
        "required": [
          "action",
          "expectedRevision"
        ]
      },
...
      "TransactionAction": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/ActionPAY"
          },
          {
            "$ref": "#/components/schemas/ActionPURCHASE"
          },
          {
            "$ref": "#/components/schemas/ActionOUTOFSTOCK"
          },
          {
            "$ref": "#/components/schemas/ActionTRAVEL"
          },
          {
            "$ref": "#/components/schemas/ActionSHIP"
          },
          {
            "$ref": "#/components/schemas/ActionRECEIVE"
          },
          {
            "$ref": "#/components/schemas/ActionRECEIVEANDCONFIRM"
          },
          {
            "$ref": "#/components/schemas/ActionCONFIRM"
          },
          {
            "$ref": "#/components/schemas/ActionSETTLE"
          },
          {
            "$ref": "#/components/schemas/ActionCANCEL"
          },
          {
            "$ref": "#/components/schemas/ActionDISPUTE"
          }
        ],
        "discriminator": {
          "propertyName": "action",
          "mapping": {
            "PAY": "#/components/schemas/ActionPAY",
            "PURCHASE": "#/components/schemas/ActionPURCHASE",
            "OUT_OF_STOCK": "#/components/schemas/ActionOUTOFSTOCK",
            "TRAVEL": "#/components/schemas/ActionTRAVEL",
            "SHIP": "#/components/schemas/ActionSHIP",
            "RECEIVE": "#/components/schemas/ActionRECEIVE",
            "RECEIVE_AND_CONFIRM": "#/components/schemas/ActionRECEIVEANDCONFIRM",
            "CONFIRM": "#/components/schemas/ActionCONFIRM",
            "SETTLE": "#/components/schemas/ActionSETTLE",
            "CANCEL": "#/components/schemas/ActionCANCEL",
            "DISPUTE": "#/components/schemas/ActionDISPUTE"
          }
        }
      }
    }
  }
}